% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuality.R
\name{getQuality}
\alias{getQuality}
\title{Gets measure of cluster quality based on Markovian summaries of ordinal sequences}
\usage{
getQuality(x)
}
\arguments{
\item{x}{an object of class \code{markovSummary} returned by the \code{markovSummary()} function}
}
\value{
a named vector containing the distinctiveness, deviation and minimum group size within the clusters
}
\description{
Gets measure of cluster quality based on Markovian summaries of ordinal sequences
}
\details{
This function returns several measures of cluster quality developed for
Ordinal Edit Distance. As the distance measure does not obey the triangle
inequality, measures of average between- and within-cluster distance that would typically be
used to measure cluster quality may be misleading. Instead, measures of
\emph{distinctiveness} and \emph{deviation} were proposed as alternatives.

To achieve this, the Markovian summary for each cluster is treated as an
"average" value for the cluster with a known, fixed number of parameters
determined by the number of levels in the ordinal sequences.
\emph{distinctiveness} is defined as the average distance between any two
Markov summaries, where distance is given by the sum of squared differences
between each parameter value. \emph{deviation} is the average error of the
Markovian summary (see \link{markovSummary}) across clusters.
}
\references{
Johns H, Hearne J, Bernhardt J, Churilov L. Clustering clinical and health
care processes using a novel measure of dissimilarity for variable-length
sequences of ordinal states.
\emph{Statistical Methods in Medical Research}.
2020;29(10):3059-3075. doi:10.1177/0962280220917174
}
